requirements.txt
google-adk>=0.3.0
google-generativeai
requests
uvicorn
